dom/imptests/webapps/DOMCore/tests/submissions/Ms2ger/test_Document-adoptNode.html

<!doctype html>
<html>
<head>
<title>Document.adoptNode</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<!--creates an element with local name "x<": --><x<>x</x<>
<script>
test(function() {
  var y = document.getElementsByTagName("x<")[0]
  assert_equals(y.parentNode, document.body)
  assert_equals(y.ownerDocument, document)
  document.adoptNode(y)
  assert_equals(y.parentNode, null)
  assert_equals(y.ownerDocument, document)
  var doc = document.implementation.createDocument(null, null, null)
  doc.adoptNode(y)
  assert_equals(y.parentNode, null)
  assert_equals(y.ownerDocument, doc)
}, "Adopting an Element called 'x<' should work.")
test(function() {
  var x = document.createElement(":good:times:")
  document.adoptNode(x)
  var doc = document.implementation.createDocument(null, null, null)
  doc.adoptNode(x)
  assert_equals(x.parentNode, null)
  assert_equals(x.ownerDocument, doc)
}, "Adopting an Element called ':good:times:' should work.")
test(function() {
  var doctype = document.doctype;
  assert_equals(doctype.parentNode, document)
  assert_equals(doctype.ownerDocument, document)
  document.adoptNode(doctype)
  assert_equals(doctype.parentNode, null)
  assert_equals(doctype.ownerDocument, document)
}, "Explicitly adopting a DocumentType should work.")
test(function() {
  var doc = document.implementation.createDocument(null, null, null)
  assert_throws("NOT_SUPPORTED_ERR", function() { document.adoptNode(doc) })
}, "Adopting a Document should throw.")
</script>
</body>
</html>
